新增角色定義

+++ api/app/service/User.js
@@ -1,12 +1,18 @@
 'use strict'
 const userRepo = App.make('Repositories/User')
+const UserCodes = use('ApiCodes/User1000')
 
-class UserService {
+class UserService
+{
   async login({request, auth}) {
     const {userName, password} = request.all()
     const tokenData = await auth.attempt(userName, password)
     const user = await userRepo.findUserByUserName(userName)
 
+    if (user.expire_time && moment(user.expire_time).diff(moment(), 'secs') < 0)
+    {
+      throw new ApiErrorException(UserCodes.USER_EXPIRED)
+    }
 
     // 新登入需刪棄用舊的token // 若是需要多處登入，則移除這行
     await userRepo.deleteOldTokensByUser(user)
